Refactor abstract alerts into util class to later roll into LEDs
